+++ bartons/src/lib.rs
@@ -22,6 +22,8 @@ fn python_kernels(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(kernels::atr::atr_py, m)?)?;
     m.add_function(wrap_pyfunction!(kernels::mad::mad_py, m)?)?;
     m.add_function(wrap_pyfunction!(kernels::cci::cci_py, m)?)?;
+    m.add_function(wrap_pyfunction!(kernels::ker::ker_py, m)?)?;
+    m.add_function(wrap_pyfunction!(kernels::kama::kama_py, m)?)?;
     m.add_function(wrap_pyfunction!(samples::random_prices_py, m)?)?;
     m.add_function(wrap_pyfunction!(samples::with_n_chunks_py, m)?)?;
     Ok(())